Move _EverestWorkflow code into a baserunmodel
